sensors/requirements.txt
PyYAML==6.0
requests==2.28.2
numpy==1.24.3
